+++ packages/ui/src/lib/sideIconSidebar/SideIconSidebarTabLabel.svelte
